Fix: 플레이리스트에 이미지 수정 시 원래 이미지 널값 검증

+++ src/main/java/crush/myList/domain/playlist/service/PlaylistService.java
@@ -80,7 +80,9 @@ public class PlaylistService {
         MultipartFile imageFile = patchRequest.getTitleImage();
 
         if (imageFile != null && !imageFile.isEmpty()) {
-            imageService.deleteImageToGcs(playlist.getImage().getId());
+            if (playlist.getImage() != null) {
+                imageService.deleteImageToGcs(playlist.getImage().getId());
+            }
             Image newImage = imageService.saveImageToGcs_Image(imageFile);
 
             playlist.setImage(newImage);
